requirements.txt
opencv-python==4.8.0.76
tensorflow==2.13.0
keras==2.13.0
numpy==1.24.3
pygame==2.2.1
Pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
